[uavcan] Rewrite to seperate module for actuators and sensors
