+++ conf/modules/actuators_uavcan.xml
@@ -3,24 +3,17 @@
 <module name="actuators_uavcan" dir="actuators" task="actuators">
   <doc>
     <description>
-      Actuators Driver for x-vert vtol escs
+      UAVCan actuators
    </description>
-    <!--configure name="ESCS_PORT" value="usb_serial" description="Port used to connect to the atmegas escs."/>
-    <configure name="ESCS_BAUD" value="250000" description="Baud rate for ESCS_PORT if a UART"/-->
   </doc>
+  <dep>
+    <depends>uavcan</depends>
+  </dep>
   <header>
     <file name="actuators_uavcan.h" dir="subsystems/actuators"/>
   </header>
-  <makefile target="ap">
-    <define name="ACTUATORS"/> 
-    <define name="USE_CAN1"/>
-    <define name="USE_CAN2"/>   
-    <file_arch name="actuators_uavcan_arch.c" dir="subsystems/actuators"/>
-    <file name="canard.c" dir="$(PAPARAZZI_SRC)/sw/ext/libcanard"/>
-    <include name="$(PAPARAZZI_SRC)/sw/ext/libcanard"/>
-  </makefile>
-  <makefile target="nps">
-    <file_arch name="actuators_uavcan_arch.c" dir="subsystems/actuators"/>
+  <makefile target="ap|nps">
+    <file name="actuators_uavcan.c" dir="subsystems/actuators"/>
   </makefile>
 </module>
