+++ conf/modules/uavcan.xml
@@ -1,0 +1,32 @@
+<!DOCTYPE module SYSTEM "module.dtd">
+
+<module name="uavcan" dir="uavcan">
+  <doc>
+    <description>
+      UAVCan interface for transmitting/receiving uavcan messages on CAN busses
+   </description>
+    <configure name="UAVCAN_USE_CAN1" value="FALSE" description="Enable UAVCAN on CAN1 interface"/>
+    <configure name="UAVCAN_USE_CAN2" value="FALSE" description="Enable UAVCAN on CAN2 interface"/>
+    <define name="UAVCAN_NODE_ID" value="100" description="The UAVCAN node ID of the AP"/>
+    <define name="UAVCAN_BAUDRATE" value="1000000" description="The baudrate of the CAN interface"/>
+  </doc>
+  <header>
+    <file name="uavcan.h" dir="modules/uavcan"/>
+  </header>
+  <init fun="uavcan_init()"/>
+  <makefile target="ap">
+    <!-- Enable the CAN busses if needed -->
+    <define name="USE_CAN1" cond="ifeq ($(UAVCAN_USE_CAN1), TRUE)"/>
+    <define name="USE_CAN2" cond="ifeq ($(UAVCAN_USE_CAN2), TRUE)"/>
+    <define name="UAVCAN_USE_CAN1" value="$(UAVCAN_USE_CAN1)"/>
+    <define name="UAVCAN_USE_CAN2" value="$(UAVCAN_USE_CAN2)"/>
+
+    <!-- Load canard -->
+    <include name="$(PAPARAZZI_SRC)/sw/ext/libcanard"/>
+    <file name="canard.c" dir="$(PAPARAZZI_SRC)/sw/ext/libcanard"/>
+
+    <!-- Load uavcan itself -->
+    <file_arch name="uavcan.c" dir="modules/uavcan"/>
+  </makefile>
+</module>
+
