+++ sw/airborne/subsystems/actuators/actuators_uavcan2.h
@@ -19,14 +19,16 @@
  * Boston, MA 02111-1307, USA.
  */
 
-#ifndef ACTUATORS_UAVCAN_H
-#define ACTUATORS_UAVCAN_H
+#ifndef ACTUATORS_UAVCAN2_H
+#define ACTUATORS_UAVCAN2_H
 
-#include "modules/uavcan/uavcan.h"
-#include BOARD_CONFIG
+#include "actuators_uavcan.h"
 
-/* External functions */
-extern void actuators_uavcan_init(struct uavcan_iface_t* iface);
-extern void actuators_uavcan_commit(struct uavcan_iface_t* iface, int16_t *values, uint8_t nb);
+/** Stub file needed per interface because of generator */
+extern int16_t actuators_uavcan2_values[SERVOS_UAVCAN2_NB];
 
-#endif /* ACTUATORS_UAVCAN_H */
+#define ActuatorsUavcan2Init() actuators_uavcan_init(&uavcan2)
+#define ActuatorUavcan2Set(_i, _v) { actuators_uavcan2_values[_i] = _v; }
+#define ActuatorsUavcan2Commit()  actuators_uavcan_commit(&uavcan2, actuators_uavcan2_values, SERVOS_UAVCAN2_NB)
+
+#endif /* ACTUATORS_UAVCAN2_H */
